init: enable language select for widgets
