+++ src/components/home/Widgets/WidgetWeather/constants.js
@@ -1,3 +1,24 @@
 const OWM_API = '46a8d7bc7f3c4adccd8efc07bf1a0431';
 
+export const labels = {
+  'en-US': {
+    feelsLike: 'Feels like',
+    wind: 'Wind',
+    windSpeed: 'm/s',
+    humidity: 'Humidity',
+  },
+  'fr-FR': {
+    feelsLike: 'Ressenti',
+    wind: 'Vent',
+    windSpeed: 'm/s',
+    humidity: 'Humidité',
+  },
+  'ru-RU': {
+    feelsLike: 'Ощущается как',
+    wind: 'Ветер',
+    windSpeed: 'м/с',
+    humidity: 'Влажность',
+  },
+};
+
 export default OWM_API;
